The issue occurs because leveldb does not support values size bigger than `1ull << 32`. The snappy compression library does not souport data larger than `1ull << 32` as well, so my new test use the option `kNoCompression`.

+++ table/block_builder.h
@@ -44,7 +44,7 @@ class BlockBuilder {
  private:
   const Options* options_;
   std::string buffer_;              // Destination buffer
-  std::vector<uint32_t> restarts_;  // Restart points
+  std::vector<uint64_t> restarts_;  // Restart points
   int counter_;                     // Number of entries emitted since restart
   bool finished_;                   // Has Finish() been called?
   std::string last_key_;
